exercise 4.21

+++ part4/blogList/controllers/blogs.js
@@ -51,6 +51,16 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (!(user && blog.user && user.id.toString() === blog.user.toString())) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
